requirements/tox.txt
psycopg2==2.6
django-braces==1.8.1
django-password-reset==0.8.1
django-bootstrap3==6.2.2
raven==5.8.1
django-admirarchy==0.2.0
django-custom-user==0.5
django-simple-menu==1.1.0
xlrd==0.9.4
arrow==0.7.0
djangorestframework==3.3.1
markdown==2.6.4
django-filter==0.11.0
celery==3.1.19
redis==2.10.5
openpyxl==2.2.5
gunicorn==19.3.0
django-jenkins
coverage
django-webtest
mock
django-nose
nose-timer